requirements.txt
numpy==1.21.2
Pillow==8.3.2
webcolors==1.11.1
